Switched to pyproject.toml and updated structure.
