+++ LICENSE
@@ -1,6 +1,7 @@
 MIT License
 
-Copyright (c) [year] [fullname]
+Copyright (c) 2020 nbgallery team
+Copyright (c) 2023 Spectric Labs, Inc.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
