--- tox.ini
@@ -1,11 +1,0 @@
-[tox]
-envlist = py{27,35,37}
-
-[testenv]
-recreate = true
-changedir = tests
-deps = discover
-commands = discover
-           python -m ipydeps six,geohash
-           python -m ipydeps urllib exifread
-           python -m ipydeps requests_toolbelt requests-toolbelt
